+++ src/main/java/uk/gov/companieshouse/acsp/models/filing/PersonName.java
@@ -1,0 +1,39 @@
+package uk.gov.companieshouse.acsp.models.filing;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class PersonName {
+
+    @JsonProperty("first_name")
+    private String firstName;
+    @JsonProperty("last_name")
+    private String lastName;
+    @JsonProperty("middle_name")
+    private String middleName;
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getMiddleName() {
+        return middleName;
+    }
+
+    public void setMiddleName(String middleName) {
+        this.middleName = middleName;
+    }
+
+
+}
